+++ spec/spec_helper.rb
@@ -17,5 +17,5 @@ RSpec.configure do |config|
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
   #     --seed 1234
-  config.order = 'default'
+  # config.order = 'default'
 end
